requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.11
pgvector==0.2.5
python-dotenv==1.0.1
langchain==0.1.12
langchain-community==0.0.28
deep-translator
sentence-transformers
